+++ vmr-newbackend/src/main/java/com/anhvan/vmr/grpc/AuthInterceptor.java
@@ -32,7 +32,7 @@ public class AuthInterceptor implements ServerInterceptor {
     }
 
     try {
-      int id = jwtUtil.authenticateBlocking(token);
+      long id = jwtUtil.authenticateBlocking(token);
       log.debug("User id get from jwt token: {}", id);
     } catch (Exception e) {
       log.debug("Fail to authenticate jwt token", e);
